+++ hardware_interface_testing/test/test_resource_manager.cpp
@@ -14,9 +14,13 @@
 
 // Authors: Karsten Knese, Denis Stogl
 
+#ifndef _USE_MATH_DEFINES
+#define _USE_MATH_DEFINES
+#endif
 #include "test_resource_manager.hpp"
 
 #include <algorithm>
+#include <cmath>
 #include <memory>
 #include <string>
 #include <unordered_map>